* commit '350d5877622482b27a18ab214dccce0fa2f1bc79':
  KERNEL-3452 1、无用代码移除、复杂代码整理重写 2、数据库相关重写

+++ swift-realtime/src/test/java/com/fr/swift/segment/backup/BackupBlockImporterTest.java
@@ -80,6 +80,6 @@ public class BackupBlockImporterTest {
         // verify new Inserting
         assertEquals(proxyInserter, Whitebox.getInternalState(inserting, "inserter"));
         assertEquals(backupSeg, Whitebox.getInternalState(inserting, "seg"));
-        assertEquals(1, Whitebox.getInternalState(inserting, "rowCount"));
+        assertEquals(1, (int) Whitebox.getInternalState(inserting, "rowCount"));
     }
 }
